Fix OverlayOutputStream resizing
Proxies 'resize' event to its listeners, so that run command can send
new dimensions to the container.

Also fixes up clearing of the overlay to prevent almost all garbage and
extraneous newlines that came up when resizing.

Uses our fork of node-charm for fixed character deletion and addition of
scrolling.