Remove our connection configuration code
The _.merge in DockerConfig was converting Buffers containing the PEM
data into arrays of bytes. This broke the TLS client authentication in
Node 4+.

Removing the code completely, since it dates to when docker-modem didn't
support getting this info on its own. Since it can now pull the daemon
auth information out of the environment, we'll just rely on it to do so.