Fix #40: recreate broken links on docker >= 1.10
Docker 1.10 changed the behavior of links when a container dies, see
https://github.com/docker/docker/issues/21049

This PR extends the logic in isLinkMissing to check if a linked
container is newer than the linking container. If it is it couldn't
possibly be actually linked since links can only be set at container
creation time. This heuristic is used to recreate containers that
crashed or stopped (but were still included as present in another
containers links) preventing the errors reported in #40.